+++ ios/Classes/CleverTapPlugin.m
@@ -480,8 +480,7 @@ static NSDateFormatter *dateFormatter;
 }
 
 - (void)profileAddMultiValues:(FlutterMethodCall *)call withResult:(FlutterResult)result {
-    
-    [[CleverTap sharedInstance] profileSetMultiValues:call.arguments[@"values"] forKey:call.arguments[@"key"]];
+    [[CleverTap sharedInstance] profileAddMultiValues:call.arguments[@"values"] forKey:call.arguments[@"key"]];
     result(nil);
 }
 
@@ -1006,6 +1005,11 @@ static NSDateFormatter *dateFormatter;
     
     [[NSNotificationCenter defaultCenter] addObserver:self
                                              selector:@selector(emitEventInternal:)
+                                                 name:kCleverTapInboxMessageTapped
+                                               object:nil];
+    
+    [[NSNotificationCenter defaultCenter] addObserver:self
+                                             selector:@selector(emitEventInternal:)
                                                  name:kCleverTapProductConfigFetched
                                                object:nil];
     
@@ -1084,14 +1088,20 @@ static NSDateFormatter *dateFormatter;
 #pragma mark CleverTapInboxViewControllerDelegate
 
 - (void)messageButtonTappedWithCustomExtras:(NSDictionary *_Nullable)customExtras {
-    
     NSMutableDictionary *body = [NSMutableDictionary new];
     if (customExtras != nil) {
-        body[@"customExtras"] = customExtras;
+        body = [NSMutableDictionary dictionaryWithDictionary:customExtras];
     }
-    [self postNotificationWithName:kCleverTapInboxMessageButtonTapped andBody:customExtras];
+    [self postNotificationWithName:kCleverTapInboxMessageButtonTapped andBody:body];
 }
 
+- (void)messageDidSelect:(CleverTapInboxMessage *_Nonnull)message atIndex:(int)index withButtonIndex:(int)buttonIndex {
+    NSMutableDictionary *body = [NSMutableDictionary new];
+    if ([message json] != nil) {
+        body = [NSMutableDictionary dictionaryWithDictionary:[message json]];
+    }
+    [self postNotificationWithName:kCleverTapInboxMessageTapped andBody:body];
+}
 
 #pragma mark CleverTapPushNotificationDelegate